Switch database path from app_paht to base_path.
This is a duplicate commit to another pull request but the code won't
work if we don't change it.
Essentially the `database` folder is no longer under `/app`. It's been
move up to base path in the Laravel 5. We could use `database_path()` or
`base_path()` depending on how we want the config file to look like.
This is for another occasion to debate.